+++ demo/src/main/java/com/aurelhubert/ahbottomnavigation/demo/DemoFragment.java
@@ -53,6 +53,7 @@ public class DemoFragment extends Fragment {
 		final DemoActivity demoActivity = (DemoActivity) getActivity();
 		final SwitchCompat switchColored = (SwitchCompat) view.findViewById(R.id.fragment_demo_switch_colored);
 		final SwitchCompat switchFiveItems = (SwitchCompat) view.findViewById(R.id.fragment_demo_switch_five_items);
+		final SwitchCompat showHideBottomNavigation = (SwitchCompat) view.findViewById(R.id.fragment_demo_show_hide);
 
 		switchColored.setChecked(demoActivity.isBottomNavigationColored());
 		switchFiveItems.setChecked(demoActivity.getBottomNavigationNbItems() == 5);
@@ -67,6 +68,12 @@ public class DemoFragment extends Fragment {
 			@Override
 			public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
 				demoActivity.updateBottomNavigationItems(isChecked);
+			}
+		});
+		showHideBottomNavigation.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+			@Override
+			public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+				demoActivity.showOrHideBottomNavigation(isChecked);
 			}
 		});
 	}